+++ CHANGELOG.md
@@ -10,9 +10,10 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ### Changed
 
-- `Equinox.SqlStreamStore.MsSql`: Update `System.IdentityModel.Tokens.Jwt` and `Microsoft.Data.SqlClient` to remove transitive vulnerable defaults [#489](https://github.com/jet/equinox/pull/489)
-- `Equinox.CosmosStore`: Update `Microsoft.Azure.Cosmos` to remove transitive vulnerable default [#489](https://github.com/jet/equinox/pull/489)
+- .NET 10 SDK update & build optimizations; Docker image update to MSSQL 2025
+- `Equinox.CosmosStore`: Update `Microsoft.Azure.Cosmos` to >= `3.58.0` to remove transitive vulnerable default [#489](https://github.com/jet/equinox/pull/489)
 - `Equinox.DynamoStore`: Update `AWSSDK` dependencies to 4.x (knock-on effect of updating `FSharp.AWS.DynamoDB` dependency to `0.13.0-beta`) [#480](https://github.com/jet/equinox/pull/480) :pray: [@njlr](https://github.com/njlr)
+- `Equinox.SqlStreamStore.MsSql`: Update `System.IdentityModel.Tokens.Jwt` to remove transitive vulnerable defaults [#489](https://github.com/jet/equinox/pull/489)
 
 ### Added